refactor!: reset Go backend and FFI model (#256)

+++ compiler/parse/variables_test.go
@@ -11,7 +11,7 @@ func TestVariables(t *testing.T) {
 			input: `
 				let name: Str = "Alice"
     		mut age: Int = 30
-        mut temp: Float = 98.6
+        mut temp: Float64 = 98.6
       	let is_student: Bool = true`,
 			output: Program{
 				Imports: []Import{},
@@ -107,7 +107,6 @@ func TestVariables(t *testing.T) {
 
 	runTests(t, tests)
 }
-
 func TestFunctionTypes(t *testing.T) {
 	runTests(t, []test{
 		// Function type error cases
@@ -219,7 +218,6 @@ func TestFunctionTypes(t *testing.T) {
 		},
 	})
 }
-
 func TestGenericCallTypeArgumentDiagnostics(t *testing.T) {
 	runTests(t, []test{
 		{
@@ -269,7 +267,6 @@ func TestGenericCallTypeArgumentDiagnostics(t *testing.T) {
 		},
 	})
 }
-
 func TestArrayMapTypes(t *testing.T) {
 	runTests(t, []test{
 		// Array/Map type error cases
@@ -305,7 +302,6 @@ func TestArrayMapTypes(t *testing.T) {
 		},
 	})
 }
-
 func TestGenericTypeParameters(t *testing.T) {
 	runTests(t, []test{
 		{
